Made tests less strict
The TypeScript compiler's purpose is only to strip type symbols and emit JavaScript that's as similar to the TS source it came from.

RSTSC is also meant to be used as a packer/bundler, so it doesn't preserve things like unnecessary parenthesis, while TSC does.

Testing becomes harder due to this, so it's been made less strict.